+++ test/td_auth_web/controllers/session_controller_test.exs
@@ -7,13 +7,20 @@ defmodule TdAuthWeb.SessionControllerTest do
   alias TdAuth.Accounts
   alias TdAuth.Auth.Guardian
   alias TdAuthWeb.ApiServices.MockAuth0Service
+  alias TdCache.Redix
+  alias TdCache.Redix.Stream
   alias TdCache.TaxonomyCache
+
+  @stream TdCache.Audit.stream()
 
   @create_attrs %{password: "temporal", user_name: "usuariotemporal", email: "some@email.com"}
   @valid_attrs %{password: "temporal", user_name: "usuariotemporal"}
   @invalid_attrs %{password: "invalido", user_name: "usuariotemporal"}
 
   setup_all do
+    on_exit(fn ->
+      Redix.del!(@stream)
+    end)
     start_supervised!(TdAuth.Accounts.UserLoader)
     start_supervised!(MockAuth0Service)
     :ok
@@ -27,6 +34,7 @@ defmodule TdAuthWeb.SessionControllerTest do
     setup [:create_user]
 
     test "create valid user session", %{conn: conn, swagger_schema: schema} do
+      Redix.del!(@stream)
       assert conn
              |> post(Routes.session_path(conn, :create),
                access_method: "access_method",
@@ -34,9 +42,44 @@ defmodule TdAuthWeb.SessionControllerTest do
              )
              |> validate_resp_schema(schema, "Token")
              |> response(:created)
+
+      assert {:ok, [event_attempt, event_success]} =
+               Stream.read(:redix, @stream, transform: true)
+
+      assert %{
+        event: "login_attempt",
+        payload: payload_attempt,
+        resource_id: "",
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ""
+      } = event_attempt
+
+      assert %{
+        "access_method" => "alternative_login",
+        "user_name" => "usuariotemporal"
+      } = Jason.decode!(payload_attempt)
+      user_id = event_success.user_id
+
+      assert %{
+        event: "login_success",
+        payload: payload_success,
+        resource_id: ^user_id,
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ^user_id
+      } = event_success
+
+      assert %{
+        "access_method" => "alternative_login",
+        "user_name" => "usuariotemporal",
+      } = Jason.decode!(payload_success)
     end
 
     test "create session with claims", %{conn: conn, swagger_schema: schema, user: user} do
+      Redix.del!(@stream)
       permission_fixture(user)
 
       assert %{"token" => token} =
@@ -52,15 +95,68 @@ defmodule TdAuthWeb.SessionControllerTest do
 
       assert %{"role" => "user"} = claims
       assert_lists_equal(claims["groups"], ["create_pg", "view_pg", "update_pg"])
+
+      assert {:ok, [event_attempt, event_success]} =
+               Stream.read(:redix, @stream, transform: true)
+
+      assert %{
+        event: "login_attempt",
+        payload: payload_attempt,
+        resource_id: "",
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ""
+      } = event_attempt
+
+      assert %{
+        "access_method" => "alternative_login",
+        "user_name" => "usuariotemporal",
+      } = Jason.decode!(payload_attempt)
+      user_id = event_success.user_id
+
+      assert %{
+        event: "login_success",
+        payload: payload_success,
+        resource_id: ^user_id,
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ^user_id
+      } = event_success
+
+      assert %{
+        "access_method" => "alternative_login",
+        "user_name" => "usuariotemporal",
+      } = Jason.decode!(payload_success)
     end
 
     test "create invalid user session", %{conn: conn} do
+      Redix.del!(@stream)
       assert conn
              |> post(Routes.session_path(conn, :create),
                access_method: "access_method",
                user: @invalid_attrs
              )
              |> response(:unauthorized)
+
+      assert {:ok, [event]} =
+               Stream.read(:redix, @stream, transform: true)
+
+      assert %{
+        event: "login_attempt",
+        payload: payload,
+        resource_id: "",
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ""
+      } = event
+
+      assert %{
+        "access_method" => "alternative_login",
+        "user_name" => "usuariotemporal"
+      } = Jason.decode!(payload)
     end
   end
 
@@ -68,6 +164,7 @@ defmodule TdAuthWeb.SessionControllerTest do
     setup [:create_user]
 
     test "create valid non existing user session", %{conn: conn, swagger_schema: schema} do
+      Redix.del!(@stream)
       {:ok, jwt, _full_claims} = Guardian.encode_and_sign(nil)
       conn = put_auth_headers(conn, jwt)
 
@@ -89,14 +186,35 @@ defmodule TdAuthWeb.SessionControllerTest do
       assert user
       assert user.full_name == Enum.join([profile[:name], profile[:family_name]], " ")
       assert user.email == profile[:email]
+
+      assert {:ok, [event]} =
+               Stream.read(:redix, @stream, transform: true)
+
+      user_id = event.user_id
+
+      assert %{
+        event: "login_success",
+        payload: payload,
+        resource_id: ^user_id,
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ^user_id
+      } = event
+
+      assert %{
+        "access_method" => nil,
+        "user_name" => "user_name"
+      } = Jason.decode!(payload)
     end
 
     test "create valid existing user session", %{conn: conn, swagger_schema: schema} do
+      Redix.del!(@stream)
       {:ok, jwt, _full_claims} = Guardian.encode_and_sign(nil)
       conn = put_auth_headers(conn, jwt)
 
       profile = %{
-        nickname: "usueariotemporal",
+        nickname: "usuariotemporal",
         name: "Un nombre especial",
         family_name: "surname",
         email: "email@especial.com"
@@ -113,9 +231,30 @@ defmodule TdAuthWeb.SessionControllerTest do
       assert user
       assert user.full_name == Enum.join([profile[:name], profile[:family_name]], " ")
       assert user.email == profile[:email]
+
+      assert {:ok, [event]} =
+               Stream.read(:redix, @stream, transform: true)
+      user_id = event.user_id
+
+      assert %{
+        event: "login_success",
+        id: _id,
+        payload: payload,
+        resource_id: ^user_id,
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ^user_id
+      } = event
+
+      assert %{
+        "access_method" => nil,
+        "user_name" => "usuariotemporal"
+      } = Jason.decode!(payload)
     end
 
     test "create invalid user session with access token", %{conn: conn} do
+      Redix.del!(@stream)
       {:ok, jwt, _full_claims} = Guardian.encode_and_sign(nil)
       profile = %{nickname: "user_name", name: "name", email: "email@xyz.com"}
       MockAuth0Service.set_user_info(401, Jason.encode!(profile))
@@ -126,9 +265,13 @@ defmodule TdAuthWeb.SessionControllerTest do
              |> response(:unauthorized)
 
       refute Accounts.get_user_by_name(profile[:nickname])
+
+      assert {:ok, []} =
+               Stream.read(:redix, @stream, transform: true)
     end
 
     test "create session proxy login when not allowed", %{conn: conn} do
+      Redix.del!(@stream)
       Application.put_env(:td_auth, :allow_proxy_login, "false")
 
       assert %{"errors" => errors} =
@@ -138,9 +281,13 @@ defmodule TdAuthWeb.SessionControllerTest do
                |> json_response(:unauthorized)
 
       assert %{"code" => "proxy_login_disabled"} = errors
+
+      assert {:ok, []} =
+               Stream.read(:redix, @stream, transform: true)
     end
 
     test "create session proxy login when is allowed and user is invalid", %{conn: conn} do
+      Redix.del!(@stream)
       Application.put_env(:td_auth, :allow_proxy_login, "true")
 
       assert %{"errors" => errors} =
@@ -150,15 +297,40 @@ defmodule TdAuthWeb.SessionControllerTest do
                |> json_response(:unauthorized)
 
       assert %{"detail" => "Invalid credentials"} = errors
+
+      assert {:ok, []} =
+               Stream.read(:redix, @stream, transform: true)
     end
 
     test "create session proxy login when is allowed and user is valid", %{conn: conn} do
+      Redix.del!(@stream)
       Application.put_env(:td_auth, :allow_proxy_login, "true")
 
       assert conn
              |> put_req_header("proxy-remote-user", "usuariotemporal")
              |> post(Routes.session_path(conn, :create))
              |> json_response(:created)
+
+      assert {:ok, [event]} =
+               Stream.read(:redix, @stream, transform: true)
+
+      user_id = event.user_id
+
+      assert %{
+        event: "login_success",
+        id: _id,
+        payload: payload,
+        resource_id: ^user_id,
+        resource_type: "auth",
+        service: "td_auth",
+        ts: _ts,
+        user_id: ^user_id
+      } = event
+
+      assert %{
+        "access_method" => "proxy_login",
+        "user_name" => "usuariotemporal"
+      } = Jason.decode!(payload)
     end
   end